Add example data to qu3 job
This commit creates a qu3.post file and a qu3-example.job
that creates the A0126 user number and populates it with the
example dataset from the user's guide. To facilite dataset creation
without an extra step of building a deadstart tape, we also sysedit QU3
onto the system using DIS before submitting the qu3-example.job.