Fix NOS Idle trigger and add support for CPU1 throttle
This commit corrects the check for CPU being in the idle package.
The check should be for if ! in monitor mode since the idle package does
not run in monitor mode.

In addition, this commit adds code to the CPU1 thread loop to check for
the idle package being in CPU1 and the number of times it's seen the
idle package in CPU1 and idles based on the global settings of the
trottle.