Refactor out idle code to function calls.
idleThrottle(CpuContext *ctx, bool checkBusy);
take in CPU context and a flag for if we should check for busy PPs
and give up CPU to the host if neccessary.

idleCheckBusy();
check if any PPs are busy and return false. only called by idleThrottle

cyber.ini now accepts:
idle=0 for off 1 for on
idlecycles=int number of idle cycles seen until we trigger
idletime=int the number of usec to sleep when we trigger.
ostype=string string indicating the os type.