Improve respnsiveness of idle throttle on Windows
This commit improves the responsiveness when we use the
idle throttle on Windows by requesting a new minimum timer resolution
using timeBeginPeriod() early on during startup.  This will cause the
system to wake us up faster from Sleep() calls than if we don't request
this.

There's a discussion on the implications of using this API call at
https://randomascii.wordpress.com/2020/10/04/windows-timer-resolution-the-great-rule-change/

The gist of it is that on versions of windows before april 2020 the
global system timer resolution would be affected by the process that
requested the lowest value.  On newer versions of windows this seems to
mainly affect the calling process and its threads while not affecting
other processes on the system as much.

The end result should be the same on earlier versions of windows and
Windows 10  and above in that we will burn more CPU cycles/power because the
system has to wake up earlier to service our Sleep() calls but we should
also perform much better when throttling and still save enough CPU
cycles to be useful.