Validation of HDR1 creation/exp dates fails after y2k.
This modset fixes validation of creation and expiration dates when
requested in the FET on a LABEL statement. HDR1 is checked when the tape
file is assigned and opened and the comparison buffer being built in this case
from the FET seems to not contain the "0" character so the comparison
always fails for years after 1999 but only if validation of those fields
had been requested.

I ran into this while playing with TMS which will always specify all of
the attributes it knows about in the database including creation and
expiration dates to be checked when it issues the LABEL requests.

My solution is to simply modify the routine in 1MT overlay 3MQ to skip
over the century char and only check 5 instead of 6 chars since the
internal format NOS uses doesn't contain a century character anyway.

This does not affect the writing of century characters in the ANSI HDR1 tape
labels which was always working fine.